refactor(observability): consolidar tríade de redação em _redaction.py
Extrai `_REDACT_RE`, `_redact` e `_safe_attrs` — copiadas byte-a-byte
em `dispatch_export.py` e `dispatch_log_export.py` — para o módulo-folha
`deile/observability/_redaction.py` (stdlib pura: `re` + `typing`).

Os dois consumidores passam a importar `from ._redaction import ...`.
O import órfão `import re` / `import re as _re` foi removido de cada
módulo. `_safe_str` permanece local em `dispatch_export.py` (não estava
duplicada).

Refs #760
